Refactored file to multiple files for modularity & readability
