+++ Player Performance Analyzer/Players_Data.h
@@ -1,0 +1,235 @@
+#ifndef PLAYERS_DATA_H
+#define PLAYERS_DATA_H
+
+typedef struct {
+    int id;
+    const char* name;
+    const char* team;
+    const char* role;
+    int totalRuns;
+    float battingAverage;
+    float strikeRate;
+    int wickets;
+    float economyRate;
+} Player;
+
+static const Player players[] = {
+    {707, "Gulbadin Naib", "Afghanistan", "All-rounder", 1200, 22.5, 83.5, 65, 5.3},
+    {718, "Sharafuddin Ashraf", "Afghanistan", "All-rounder", 450, 22.5, 74.3, 40, 4.9},
+    {713, "Mujeeb Ur Rahman", "Afghanistan", "Bowler", 250, 10.9, 66.3, 95, 4.4},
+    {717, "Qais Ahmad", "Afghanistan", "Bowler", 80, 8.9, 60.2, 22, 5.2},
+    {1005, "Dasun Shanaka", "Sri Lanka", "All-rounder", 1350, 28.6, 92.4, 48, 6.1},
+    {608, "Romario Shepherd", "West Indies", "All-rounder", 780, 26.4, 98.1, 45, 6.1},
+    {306, "Ben Stokes", "England", "All-rounder", 3250, 39.1, 95.7, 80, 6.0},
+    {318, "Will Jacks", "England", "Batsman", 400, 30.0, 105.5, 5, 6.0},
+    {413, "Abbas Afridi", "Pakistan", "Bowler", 45, 9.0, 65.0, 28, 6.0},
+    {918, "Beuran Hendricks", "South Africa", "Bowler", 70, 8.5, 65.2, 25, 6.0},
+    {1012, "Chamika Karunaratne", "Sri Lanka", "All-rounder", 510, 23.1, 89.7, 42, 6.0},
+    {212, "Sean Abbott", "Australia", "All-rounder", 220, 16.8, 87.5, 65, 5.9},
+    {219, "Mitchell Swepson", "Australia", "Bowler", 60, 7.2, 60.1, 28, 5.9},
+    {415, "Haris Rauf", "Pakistan", "Bowler", 95, 9.2, 72.2, 75, 5.9},
+    {509, "Lockie Ferguson", "New Zealand", "Bowler", 180, 10.8, 70.0, 95, 5.9},
+    {908, "Marco Jansen", "South Africa", "All-rounder", 620, 28.6, 102.2, 45, 5.9},
+    {1015, "Lahiru Kumara", "Sri Lanka", "Bowler", 80, 9.8, 66.5, 45, 5.9},
+    {1215, "Ebadot Hossain", "Bangladesh", "Bowler", 50, 7.9, 60.3, 20, 5.9},
+    {714, "Fazalhaq Farooqi", "Afghanistan", "Bowler", 150, 8.6, 59.4, 60, 5.1},
+    {712, "Noor Ahmad", "Afghanistan", "Bowler", 100, 9.1, 65.2, 25, 4.8},
+    {719, "Hamid Hassan", "Afghanistan", "Bowler", 90, 6.2, 57.8, 55, 4.5},
+    {709, "Rashid Khan", "Afghanistan", "All-rounder", 1200, 19.6, 98.2, 180, 4.3},
+    {320, "George Garton", "England", "Bowler", 60, 8.0, 70.5, 22, 6.3},
+    {317, "Liam Livingstone", "England", "All-rounder", 1100, 35.0, 124.0, 35, 6.2},
+    {609, "Andre Russell", "West Indies", "All-rounder", 1034, 28.0, 130.2, 70, 6.2},
+    {109, "Shardul Thakur", "India", "All-rounder", 350, 18.5, 102.3, 70, 6.1},
+    {218, "Matthew Short", "Australia", "Batsman", 290, 28.5, 94.2, 5, 6.1},
+    {319, "Chris Jordan", "England", "Bowler", 250, 15.0, 92.0, 80, 6.1},
+    {101, "Virat Kohli", "India", "Batsman", 13848, 58.7, 93.6, 4, 5.8},
+    {120, "Deepak Chahar", "India", "All-rounder", 400, 18.8, 90.1, 65, 5.8},
+    {308, "Sam Curran", "England", "All-rounder", 720, 25.7, 92.5, 65, 5.8},
+    {411, "Usama Mir", "Pakistan", "Bowler", 80, 8.5, 68.0, 40, 5.8},
+    {1212, "Afif Hossain", "Bangladesh", "All-rounder", 920, 28.1, 89.3, 30, 5.8},
+    {520, "Ben Sears", "New Zealand", "Bowler", 70, 9.8, 68.0, 25, 5.8},
+    {909, "Andile Phehlukwayo", "South Africa", "All-rounder", 1250, 26.8, 97.4, 69, 5.8},
+    {1017, "Matheesha Pathirana", "Sri Lanka", "Bowler", 60, 8.0, 69.0, 30, 5.8},
+    {603, "Rovman Powell", "West Indies", "All-rounder", 1450, 29.5, 103.2, 28, 5.8},
+    {612, "Oshane Thomas", "West Indies", "Bowler", 110, 9.4, 70.1, 55, 5.8},
+    {206, "Glenn Maxwell", "Australia", "All-rounder", 3840, 35.1, 125.4, 70, 5.7},
+    {220, "Jason Behrendorff", "Australia", "Bowler", 80, 9.0, 64.2, 45, 5.7},
+    {314, "David Willey", "England", "All-rounder", 550, 22.0, 89.5, 90, 5.7},
+    {412, "Faheem Ashraf", "Pakistan", "All-rounder", 400, 22.0, 88.0, 70, 5.7},
+    {1211, "Mohammad Saifuddin", "Bangladesh", "All-rounder", 850, 26.4, 85.6, 65, 5.7},
+    {507, "James Neesham", "New Zealand", "All-rounder", 1465, 28.9, 96.8, 72, 5.7},
+    {910, "Wayne Parnell", "South Africa", "All-rounder", 985, 25.1, 88.3, 100, 5.7},
+    {1013, "Kasun Rajitha", "Sri Lanka", "Bowler", 90, 10.4, 70.6, 55, 5.7},
+    {615, "Keemo Paul", "West Indies", "All-rounder", 520, 24.8, 87.2, 40, 5.7},
+    {207, "Marcus Stoinis", "Australia", "All-rounder", 1612, 29.8, 95.5, 52, 5.6},
+    {106, "Hardik Pandya", "India", "All-rounder", 1764, 34.8, 115.5, 83, 5.6},
+    {216, "Nathan Ellis", "Australia", "Bowler", 75, 9.5, 72.3, 25, 5.6},
+    {310, "Adil Rashid", "England", "Bowler", 330, 13.8, 78.5, 185, 5.6},
+    {513, "Ish Sodhi", "New Zealand", "Bowler", 190, 11.0, 74.1, 85, 5.6},
+    {414, "Salman Ali Agha", "Pakistan", "All-rounder", 510, 31.2, 84.3, 25, 5.6},
+    {920, "Heinrich Mulder", "South Africa", "All-rounder", 480, 29.5, 86.3, 38, 5.6},
+    {1011, "Dunith Wellalage", "Sri Lanka", "All-rounder", 420, 25.0, 82.3, 35, 5.6},
+    {1218, "Soumya Sarkar", "Bangladesh", "All-rounder", 1760, 31.1, 93.1, 43, 5.6},
+    {606, "Kyle Mayers", "West Indies", "All-rounder", 1050, 32.2, 92.6, 35, 5.6},
+    {618, "Fabian Allen", "West Indies", "All-rounder", 510, 23.6, 101.3, 38, 5.6},
+    {118, "Prasidh Krishna", "India", "Bowler", 90, 8.5, 70.0, 50, 5.5},
+    {313, "Reece Topley", "England", "Bowler", 95, 9.8, 70.0, 55, 5.5},
+    {416, "Iftikhar Ahmed", "Pakistan", "All-rounder", 950, 37.0, 96.5, 35, 5.5},
+    {904, "Aiden Markram", "South Africa", "All-rounder", 2100, 36.4, 93.7, 32, 5.5},
+    {915, "Gerald Coetzee", "South Africa", "Bowler", 85, 11.0, 75.0, 28, 5.5},
+    {1016, "Jeffrey Vandersay", "Sri Lanka", "Bowler", 70, 8.9, 62.1, 28, 5.5},
+    {1220, "Shoriful Islam", "Bangladesh", "Bowler", 70, 8.9, 67.4, 25, 5.5},
+    {111, "Mohammed Shami", "India", "Bowler", 250, 12.8, 76.3, 204, 5.4},
+    {211, "Adam Zampa", "Australia", "Bowler", 120, 8.1, 66.0, 155, 5.4},
+    {307, "Moeen Ali", "England", "All-rounder", 2200, 26.8, 103.1, 99, 5.4},
+    {905, "David Miller", "South Africa", "Batsman", 4605, 43.0, 102.4, 0, 0.0},
+    {906, "Heinrich Klaasen", "South Africa", "Batsman", 1800, 41.2, 112.7, 0, 0.0},
+    {907, "Reeza Hendricks", "South Africa", "Batsman", 1700, 38.8, 89.2, 0, 0.0},
+    {916, "Tristan Stubbs", "South Africa", "Batsman", 360, 27.0, 98.2, 0, 0.0},
+    {917, "Ryan Rickelton", "South Africa", "Batsman", 420, 35.0, 91.7, 0, 0.0},
+    {1009, "Dimuth Karunaratne", "Sri Lanka", "Batsman", 2200, 33.7, 76.4, 0, 0.0},
+    {1018, "Avishka Fernando", "Sri Lanka", "Batsman", 1550, 34.7, 91.6, 0, 0.0},
+    {1019, "Danushka Gunathilaka", "Sri Lanka", "Batsman", 1900, 33.5, 87.1, 0, 0.0},
+    {1020, "Bhanuka Rajapaksa", "Sri Lanka", "Batsman", 880, 30.4, 98.5, 0, 0.0},
+    {1202, "Tamim Iqbal", "Bangladesh", "Batsman", 8313, 36.6, 78.5, 0, 0.0},
+    {1216, "Anamul Haque", "Bangladesh", "Batsman", 1300, 31.5, 82.9, 0, 0.0},
+    {1217, "Yasir Ali Chowdhury", "Bangladesh", "Batsman", 700, 28.4, 85.7, 0, 0.0},
+    {1219, "Mominul Haque", "Bangladesh", "Batsman", 1500, 33.2, 77.8, 0, 0.0},
+    {311, "Mark Wood", "England", "Bowler", 210, 12.0, 81.0, 120, 5.4},
+    {518, "Jimmy Neesham", "New Zealand", "All-rounder", 1500, 30.5, 95.8, 65, 5.4},
+    {506, "Glenn Phillips", "New Zealand", "All-rounder", 1050, 37.5, 98.5, 30, 5.4},
+    {511, "Tim Southee", "New Zealand", "Bowler", 300, 12.5, 72.2, 214, 5.4},
+    {912, "Lungi Ngidi", "South Africa", "Bowler", 150, 10.3, 68.7, 85, 5.4},
+    {1014, "Dushmantha Chameera", "Sri Lanka", "Bowler", 140, 11.1, 77.2, 65, 5.4},
+    {616, "Shannon Gabriel", "West Indies", "Bowler", 75, 8.5, 64.3, 33, 5.4},
+    {102, "Rohit Sharma", "India", "Batsman", 10709, 49.1, 90.2, 8, 5.3},
+    {205, "Mitchell Marsh", "Australia", "All-rounder", 2285, 34.7, 93.2, 62, 5.3},
+    {217, "Cameron Green", "Australia", "All-rounder", 640, 33.7, 88.0, 32, 5.3},
+    {514, "Rachin Ravindra", "New Zealand", "All-rounder", 640, 34.7, 91.8, 32, 5.3},
+    {410, "Hasan Ali", "Pakistan", "Bowler", 310, 13.0, 75.2, 120, 5.3},
+    {1210, "Nasir Hossain", "Bangladesh", "All-rounder", 1100, 32.5, 79.2, 45, 5.3},
+    {607, "Jason Holder", "West Indies", "All-rounder", 2150, 27.8, 88.7, 158, 5.3},
+    {119, "Washington Sundar", "India", "All-rounder", 350, 25.0, 85.2, 45, 5.2},
+    {203, "Travis Head", "Australia", "Batsman", 2700, 42.2, 99.8, 2, 5.2},
+    {309, "Chris Woakes", "England", "All-rounder", 1400, 25.1, 90.2, 160, 5.2},
+    {406, "Shadab Khan", "Pakistan", "All-rounder", 870, 25.1, 92.3, 85, 5.2},
+    {1004, "Dhananjaya de Silva", "Sri Lanka", "All-rounder", 1900, 31.2, 85.3, 45, 5.2},
+    {420, "Naseem Shah", "Pakistan", "Bowler", 75, 10.5, 68.2, 52, 5.2},
+    {1207, "Taskin Ahmed", "Bangladesh", "Bowler", 210, 10.8, 75.6, 95, 5.2},
+    {1214, "Shoriful Islam", "Bangladesh", "Bowler", 80, 9.5, 68.5, 35, 5.2},
+    {505, "Daryl Mitchell", "New Zealand", "All-rounder", 2150, 44.2, 92.7, 45, 5.2},
+    {519, "Kyle Jamieson", "New Zealand", "Bowler", 160, 12.4, 73.3, 45, 5.2},
+    {911, "Kagiso Rabada", "South Africa", "Bowler", 280, 12.4, 70.8, 157, 5.2},
+    {611, "Alzarri Joseph", "West Indies", "Bowler", 180, 10.6, 68.7, 125, 5.2},
+    {617, "Roston Chase", "West Indies", "All-rounder", 960, 28.3, 79.5, 45, 5.2},
+    {113, "Mohammed Siraj", "India", "Bowler", 140, 8.2, 65.2, 95, 5.1},
+    {208, "Pat Cummins", "Australia", "Bowler", 350, 14.2, 81.3, 140, 5.1},
+    {210, "Mitchell Starc", "Australia", "Bowler", 450, 13.6, 85.2, 236, 5.1},
+    {512, "Matt Henry", "New Zealand", "Bowler", 170, 10.0, 70.8, 135, 5.1},
+    {408, "Mohammad Nawaz", "Pakistan", "All-rounder", 750, 27.4, 90.1, 65, 5.1},
+    {1204, "Mahmudullah", "Bangladesh", "All-rounder", 4950, 35.2, 82.1, 82, 5.1},
+    {614, "Gudakesh Motie", "West Indies", "Bowler", 95, 8.9, 66.9, 42, 5.1},
+    {312, "Jofra Archer", "England", "Bowler", 180, 10.4, 85.1, 75, 5.0},
+    {913, "Tabraiz Shamsi", "South Africa", "Bowler", 100, 9.2, 60.1, 75, 5.0},
+    {112, "Kuldeep Yadav", "India", "Bowler", 185, 10.4, 72.5, 165, 4.9},
+    {114, "Ravichandran Ashwin", "India", "All-rounder", 707, 16.9, 83.2, 151, 4.9},
+    {107, "Ravindra Jadeja", "India", "All-rounder", 2750, 32.3, 87.2, 220, 4.9},
+    {407, "Imad Wasim", "Pakistan", "All-rounder", 985, 38.2, 91.5, 63, 4.9},
+    {1006, "Angelo Mathews", "Sri Lanka", "All-rounder", 5865, 41.3, 84.5, 120, 4.9},
+    {1208, "Mustafizur Rahman", "Bangladesh", "Bowler", 150, 9.2, 70.1, 160, 4.9},
+    {510, "Trent Boult", "New Zealand", "Bowler", 240, 11.3, 68.5, 211, 4.9},
+    {914, "Keshav Maharaj", "South Africa", "Bowler", 230, 15.1, 72.2, 95, 4.9},
+    {610, "Akeal Hosein", "West Indies", "Bowler", 290, 14.8, 82.4, 75, 4.9},
+    {108, "Axar Patel", "India", "All-rounder", 585, 23.4, 98.0, 65, 4.8},
+    {215, "Ashton Agar", "Australia", "All-rounder", 330, 21.2, 84.6, 55, 4.8},
+    {409, "Shaheen Afridi", "Pakistan", "Bowler", 210, 11.5, 70.4, 145, 4.8},
+    {1209, "Mehidy Hasan Miraz", "Bangladesh", "All-rounder", 1200, 29.5, 82.0, 95, 4.8},
+    {601, "Shai Hope", "West Indies", "Batsman", 5400, 51.3, 75.8, 0, 0.0},
+    {602, "Nicholas Pooran", "West Indies", "Batsman", 2200, 36.7, 95.4, 0, 0.0},
+    {604, "Shimron Hetmyer", "West Indies", "Batsman", 1900, 34.6, 97.5, 0, 0.0},
+    {605, "Brandon King", "West Indies", "Batsman", 1120, 33.8, 89.3, 0, 0.0},
+    {508, "Mitchell Santner", "New Zealand", "All-rounder", 1250, 27.3, 87.2, 95, 4.8},
+    {919, "Bjorn Fortuin", "South Africa", "Bowler", 60, 7.8, 62.3, 18, 4.8},
+    {1010, "Maheesh Theekshana", "Sri Lanka", "Bowler", 250, 12.5, 68.1, 65, 4.8},
+    {613, "Kemar Roach", "West Indies", "Bowler", 310, 12.3, 72.8, 130, 4.8},
+    {209, "Josh Hazlewood", "Australia", "Bowler", 210, 12.1, 70.2, 135, 4.7},
+    {1213, "Taijul Islam", "Bangladesh", "Bowler", 120, 8.7, 65.2, 45, 4.7},
+    {110, "Jasprit Bumrah", "India", "Bowler", 190, 12.2, 80.1, 230, 4.6},
+    {1201, "Shakib Al Hasan", "Bangladesh", "All-rounder", 7550, 38.0, 83.5, 317, 4.6},
+    {103, "Shubman Gill", "India", "Batsman", 2271, 61.4, 103.1, 0, 0.0},
+    {104, "KL Rahul", "India", "Batsman", 2820, 47.0, 88.2, 0, 0.0},
+    {116, "Suryakumar Yadav", "India", "Batsman", 950, 38.0, 102.8, 0, 0.0},
+    {701, "Rahmanullah Gurbaz", "Afghanistan", "Batsman", 1800, 40.0, 89.5, 0, 0.0},
+    {702, "Ibrahim Zadran", "Afghanistan", "Batsman", 1600, 44.4, 83.2, 0, 0.0},
+    {703, "Hashmatullah Shahidi", "Afghanistan", "Batsman", 2300, 32.8, 74.6, 0, 0.0},
+    {704, "Riaz Hassan", "Afghanistan", "Batsman", 550, 30.5, 76.2, 0, 0.0},
+    {705, "Najibullah Zadran", "Afghanistan", "Batsman", 2000, 31.2, 86.1, 0, 0.0},
+    {117, "Sanju Samson", "India", "Batsman", 800, 40.0, 95.0, 0, 0.0},
+    {201, "David Warner", "Australia", "Batsman", 6932, 45.2, 95.3, 0, 0.0},
+    {202, "Steve Smith", "Australia", "Batsman", 5054, 44.5, 87.1, 0, 0.0},
+    {115, "Ishan Kishan", "India", "Batsman", 1004, 45.6, 110.7, 0, 0.0},
+    {204, "Marnus Labuschagne", "Australia", "Batsman", 1400, 38.5, 83.6, 0, 0.0},
+    {105, "Shreyas Iyer", "India", "Batsman", 1870, 44.5, 94.8, 0, 0.0},
+    {213, "Alex Carey", "Australia", "Batsman", 1850, 36.4, 88.9, 0, 0.0},
+    {214, "Josh Inglis", "Australia", "Batsman", 480, 32.0, 91.1, 0, 0.0},
+    {301, "Joe Root", "England", "Batsman", 6520, 49.7, 86.3, 0, 0.0},
+    {302, "Jonny Bairstow", "England", "Batsman", 3930, 46.2, 101.4, 0, 0.0},
+    {710, "Karim Janat", "Afghanistan", "All-rounder", 650, 29.5, 88.1, 28, 5.7},
+    {708, "Mohammad Nabi", "Afghanistan", "All-rounder", 3200, 27.4, 86.8, 165, 4.6},
+    {716, "Naveen-ul-Haq", "Afghanistan", "Bowler", 120, 9.5, 64.8, 45, 5.7},
+    {715, "Farid Ahmad Malik", "Afghanistan", "Bowler", 100, 7.9, 61.5, 30, 5.4},
+    {706, "Azmatullah Omarzai", "Afghanistan", "All-rounder", 850, 33.9, 91.4, 35, 5.8},
+    {303, "Jos Buttler", "England", "Batsman", 4820, 41.0, 118.2, 0, 0.0},
+    {304, "Dawid Malan", "England", "Batsman", 1600, 46.5, 95.3, 0, 0.0},
+    {305, "Jason Roy", "England", "Batsman", 4300, 39.2, 104.8, 0, 0.0},
+    {315, "Phil Salt", "England", "Batsman", 650, 34.0, 112.0, 0, 0.0},
+    {316, "Harry Brook", "England", "Batsman", 800, 41.0, 99.0, 0, 0.0},
+    {401, "Babar Azam", "Pakistan", "Batsman", 5720, 57.6, 89.3, 0, 0.0},
+    {711, "Ikram Alikhil", "Afghanistan", "Batsman", 500, 25.0, 72.9, 0, 0.0},
+    {402, "Fakhar Zaman", "Pakistan", "Batsman", 3850, 45.0, 94.8, 0, 0.0},
+    {403, "Imam-ul-Haq", "Pakistan", "Batsman", 3300, 51.2, 84.7, 0, 0.0},
+    {404, "Abdullah Shafique", "Pakistan", "Batsman", 750, 43.0, 85.0, 0, 0.0},
+    {515, "Mark Chapman", "New Zealand", "Batsman", 780, 36.0, 90.5, 0, 0.0},
+    {516, "Henry Nicholls", "New Zealand", "Batsman", 1990, 36.2, 85.0, 0, 0.0},
+    {517, "Tom Blundell", "New Zealand", "Batsman", 520, 32.1, 89.1, 0, 0.0},
+    {405, "Mohammad Rizwan", "Pakistan", "Batsman", 1900, 40.2, 89.6, 0, 0.0},
+    {1001, "Kusal Mendis", "Sri Lanka", "Batsman", 3900, 36.5, 87.2, 0, 0.0},
+    {1002, "Pathum Nissanka", "Sri Lanka", "Batsman", 2300, 47.1, 90.8, 0, 0.0},
+    {1003, "Charith Asalanka", "Sri Lanka", "Batsman", 1850, 39.0, 89.9, 0, 0.0},
+    {1007, "Kusal Perera", "Sri Lanka", "Batsman", 3300, 32.2, 92.8, 0, 0.0},
+    {1008, "Sadeera Samarawickrama", "Sri Lanka", "Batsman", 900, 40.9, 88.7, 0, 0.0},
+    {417, "Sarfaraz Ahmed", "Pakistan", "Batsman", 2350, 34.1, 85.8, 0, 0.0},
+    {418, "Saim Ayub", "Pakistan", "Batsman", 210, 28.5, 94.1, 0, 0.0},
+    {419, "Shan Masood", "Pakistan", "Batsman", 1000, 32.5, 82.0, 0, 0.0},
+    {501, "Kane Williamson", "New Zealand", "Batsman", 6675, 47.8, 82.1, 0, 0.0},
+    {1203, "Mushfiqur Rahim", "Bangladesh", "Batsman", 7800, 37.4, 79.8, 0, 0.0},
+    {1205, "Liton Das", "Bangladesh", "Batsman", 2600, 33.9, 88.3, 0, 0.0},
+    {1206, "Najmul Hossain Shanto", "Bangladesh", "Batsman", 1850, 41.2, 84.9, 0, 0.0},
+    {502, "Devon Conway", "New Zealand", "Batsman", 1810, 47.6, 88.9, 0, 0.0},
+    {503, "Tom Latham", "New Zealand", "Batsman", 4330, 36.5, 86.0, 0, 0.0},
+    {504, "Will Young", "New Zealand", "Batsman", 1250, 41.2, 87.4, 0, 0.0},
+    {901, "Quinton de Kock", "South Africa", "Batsman", 6770, 45.3, 96.8, 0, 0.0},
+    {902, "Temba Bavuma", "South Africa", "Batsman", 2005, 44.0, 88.6, 0, 0.0},
+    {903, "Rassie van der Dussen", "South Africa", "Batsman", 2385, 55.0, 89.5, 0, 0.0},
+    {619, "Johnson Charles", "West Indies", "Batsman", 1080, 30.2, 90.1, 0, 0.0},
+    {620, "Evin Lewis", "West Indies", "Batsman", 1950, 36.4, 89.8, 0, 0.0},
+    {720, "Usman Ghani", "Afghanistan", "Batsman", 1050, 26.9, 77.4, 0, 0.0}
+};
+
+static const char* teams[] = {
+    "Afghanistan",
+    "Australia",
+    "Bangladesh",
+    "England",
+    "India",
+    "New Zealand",
+    "Pakistan",
+    "South Africa",
+    "Sri Lanka",
+    "West Indies"
+};
+
+static int teamCount = sizeof(teams) / sizeof(teams[0]);
+static int playerCount = sizeof(players) / sizeof(players[0]);
+
+#endif
