+++ Player Performance Analyzer/header.h
@@ -1,0 +1,96 @@
+#ifndef HEADER_H
+#define HEADER_H
+
+#include<stdio.h>
+#include<stdlib.h>
+#include<string.h>
+#include<stdbool.h>
+#include "Players_Data.h"
+
+#define NAME_SIZE 100
+#define TEAM_NAME_SIZE 100
+#define STR_BOWLER "Bowler"
+#define STR_BATSMAN "Batsman"
+#define STR_ALL_ROUNDER "All-rounder"
+#define PERFORMANCE_DIVISION 100
+#define WICKET_POINTS 2
+#define ECONOMY_BASE 100 
+#define ROLE_COUNT 3
+#define HEAP_ROOT_INDEX 0
+
+typedef enum {
+    BATSMAN = 1,
+    BOWLER = 2,
+    ALL_ROUNDER = 3
+} PlayerRole;
+
+typedef struct PlayerData {
+    int id;
+    char name[NAME_SIZE];
+    char teamName[TEAM_NAME_SIZE]; 
+    PlayerRole role;
+    int totalRuns;
+    float battingAverage;
+    float strikeRate;
+    int wickets;
+    float economyRate;
+    float performanceIndex;
+    struct PlayerData* next;
+} PlayerData;
+
+typedef struct Team {
+    int id;
+    char name[NAME_SIZE];
+    int totalPlayers;
+    int battingPlayerCount;
+    float averageBattingStrikeRate;
+    PlayerData* batsmanHead;
+    PlayerData* bowlerHead;
+    PlayerData* allRounderHead;
+} Team;
+
+typedef struct HeapNode {
+    PlayerData* player;
+    int teamId;
+} HeapNode;
+
+//Command Handlers
+void displayMenu();
+void handleAddPlayer(Team* team, int teamId);
+void handleTopKPlayers(Team* team, int teamId);
+void handleChoice(Team* team);
+
+void createTeam(Team* currentTeam, int teamId, const char* teamName);
+Team* initialiseTeams();
+void addPlayerToTeam(Team* team, PlayerData* newPlayer);
+bool createPlayer(Team* team, Player player);
+bool initialisePlayers(Team* team);
+PlayerData* createNewPlayer();
+bool addNewPlayerToTeam(Team* team, int teamId);
+bool displayTeamData(Team* team, int teamId);
+bool displaySortedTeams(Team* team);
+bool getTopKPlayers(Team* team, int teamId, PlayerRole role, int k);
+void displaySortedPlayers(Team* team, PlayerRole role);
+void displaySortedPlayersOfRole(Team* team, PlayerRole role);
+void freePlayers(PlayerData* head);
+void freeAllTeams(Team* team);
+
+PlayerRole getRoleByString(const char* role);
+void calculatePerformanceIndex(PlayerData* player);
+Team* searchTeamByName(Team* team, const char *teamName);
+PlayerData** getHeadByRole(Team *team, PlayerRole role);
+void updateAverageBattingStrikeRate(Team* team, PlayerData* player);
+Team* searchTeamById(Team* team, int teamId);
+void displayPlayerTableHeader();
+char *getStringByRole(PlayerRole role);
+void displayPlayerDetails(PlayerData* player);
+void swap(Team* team1, Team* team2);
+int partitionIndex(Team* team, int low, int high);
+void quickSort(Team* team, int low, int high);
+void initialiseHeap(Team* team, PlayerRole role, HeapNode* heap, int* heapSize);
+void maxHeapify(HeapNode* heap, int size, int index);
+void buildMaxHeap(HeapNode* heap, int heapSize);
+HeapNode extractMax(HeapNode heap[], int *heapSize);
+
+
+#endif
